Rename to bettor reflect its use
